+++ src/app/components/DownshiftInput/DownshiftInput.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback, useMemo } from 'react';
 import ReactDOM from 'react-dom';
 import Downshift from 'downshift';
-import { Mention, MentionsInput, SuggestionDataItem } from 'react-mentions';
 import { useSelector } from 'react-redux';
 import { ResolveTokenValuesResult } from '@/plugin/tokenHelpers';
 import Box from '../Box';
@@ -11,100 +10,20 @@ import Stack from '../Stack';
 import { SingleToken } from '@/types/tokens';
 import Input, { StyledPrefix } from '../Input';
 import { TokenTypes } from '@/constants/TokenTypes';
-import { styled } from '@/stitches.config';
 import Tooltip from '../Tooltip';
 import { Properties } from '@/constants/Properties';
 import { isDocumentationType } from '@/utils/is/isDocumentationType';
 import { useReferenceTokenType } from '@/app/hooks/useReferenceTokenType';
-import mentionInputStyles from './mentionInputStyle';
 import { ErrorValidation } from '../ErrorValidation';
 import useFigmaFonts from '@/hooks/useFigmaFonts';
 import { figmaFontsSelector } from '@/selectors';
+import { MentionInput } from './MentionInput';
+import {
+  StyledButton,
+  StyledDropdown, StyledItem, StyledItemColor, StyledItemColorDiv, StyledItemName, StyledItemValue, StyledPart,
+} from './StyledDownshiftInput';
 
 type SearchField = 'Tokens' | 'Fonts' | 'Weights';
-
-const StyledDropdown = styled('div', {
-  position: 'absolute',
-  zIndex: '10',
-  width: '100%',
-  maxHeight: '300px',
-  borderRadius: '$contextMenu',
-  overflowY: 'scroll',
-  backgroundColor: '$bgDefault',
-  marginTop: '1px',
-  cursor: 'pointer',
-  boxShadow: '$contextMenu',
-});
-
-const StyledItemValue = styled('div', {
-  color: '$textMuted',
-  fontWeight: '$bold',
-  textAlign: 'right',
-  textTransform: 'uppercase',
-  maxWidth: '300px',
-  overflow: 'hidden',
-  textOverflow: 'ellipsis',
-});
-
-const StyledItem = styled('div', {
-  display: 'flex',
-  alignItems: 'center',
-  gap: '$2',
-  padding: '$2 $3',
-  fontSize: '$xsmall',
-  variants: {
-    isFocused: {
-      true: {
-        backgroundColor: '$interaction',
-        color: '$onInteraction',
-        [`& ${StyledItemValue}`]: {
-          color: '$onInteraction',
-        },
-      },
-    },
-  },
-});
-
-const StyledItemColorDiv = styled('div', {
-  flexShrink: 0,
-});
-
-const StyledItemColor = styled('div', {
-  width: '16px',
-  height: '16px',
-  borderRadius: '$colorSwatch',
-  border: '1px solid',
-  borderColor: '$borderMuted',
-});
-
-const StyledItemName = styled('div', {
-  flexGrow: 1,
-  overflow: 'hidden',
-  textOverflow: 'ellipsis',
-});
-
-const StyledPart = styled('span', {
-  variants: {
-    matches: {
-      true: {
-        fontWeight: '$bold',
-      },
-    },
-  },
-});
-
-const StyledButton = styled('button', {
-  padding: '$2 $3',
-  fontSize: '$xsmall',
-
-  variants: {
-    isFocused: {
-      false: {
-        color: '$textDisabled',
-      },
-    },
-  },
-});
 
 interface DownShiftProps {
   name?: string;
@@ -221,25 +140,6 @@ export const DownshiftInput: React.FunctionComponent<DownShiftProps> = ({
     return [];
   }, [figmaFonts, currentSearchField, searchInput, externalFontFamily]);
 
-  const mentionData = useMemo<SuggestionDataItem[]>(() => {
-    if (isDocumentationType(type as Properties)) {
-      return resolvedTokens
-        .filter((token: SingleToken) => token.name !== initialName).sort((a, b) => (
-          a.name.localeCompare(b.name, undefined, { numeric: true, sensitivity: 'base' })
-        )).map((resolvedToken) => ({
-          id: resolvedToken.name,
-          display: resolvedToken.name,
-        }));
-    }
-    return resolvedTokens
-      .filter((token: SingleToken) => referenceTokenTypes.includes(token?.type) && token.name !== initialName).sort((a, b) => (
-        a.name.localeCompare(b.name, undefined, { numeric: true, sensitivity: 'base' })
-      )).map((resolvedToken) => ({
-        id: resolvedToken.name,
-        display: resolvedToken.name,
-      }));
-  }, [initialName, resolvedTokens, referenceTokenTypes, type]);
-
   const getHighlightedText = useCallback((text: string, highlight: string) => {
     // Split on highlight term and include term into parts, ignore case
     const parts = text.split(new RegExp(`(${highlight.replace(/[-/\\^$*+?.()|[\]{}]/g, '\\$&')})`, 'gi'));
@@ -294,17 +194,6 @@ export const DownshiftInput: React.FunctionComponent<DownShiftProps> = ({
     setShowAutoSuggest(!showAutoSuggest);
   }, [showAutoSuggest]);
 
-  const handleInputBlur = React.useCallback(() => {
-    if (handleBlur) {
-      handleBlur();
-    }
-  }, [handleBlur]);
-
-  const handleMentionInputChange = React.useCallback((e) => {
-    e.target.name = name ?? 'value';
-    handleChange(e);
-  }, [handleChange, name]);
-
   const handleSearchInputChange = React.useCallback<React.ChangeEventHandler<HTMLInputElement>>((e) => {
     setSearchInput(e.target.value);
   }, []);
@@ -313,48 +202,6 @@ export const DownshiftInput: React.FunctionComponent<DownShiftProps> = ({
     if (currentSearchField === 'Tokens' && externalSearchField) setCurrentSearchField(externalSearchField);
     else setCurrentSearchField('Tokens');
   }, [currentSearchField, externalSearchField]);
-
-  const renderMentionList = React.useCallback((children: React.ReactNode) => ReactDOM.createPortal(
-    <Box css={{
-      position: 'absolute', background: '$bgDefault', top: '0', width: `${inputContainerWith}px`, padding: '$2', zIndex: '10', transform: `translate(${inputContainerPosX}px, ${inputContainerPosY}px)`,
-    }}
-    >
-      <StyledDropdown className="content scroll-container">
-        {children}
-      </StyledDropdown>
-    </Box>,
-    portalPlaceholder,
-  ), [inputContainerWith, inputContainerPosX, inputContainerPosY, portalPlaceholder]);
-
-  const renderMentionListItem = React.useCallback((
-    suggestion: SuggestionDataItem,
-    search,
-    highlightedDisplay,
-    focused,
-  ) => {
-    const resolvedToken = resolvedTokens.find((token) => token.type === type && token.name === suggestion.id);
-    return (
-      <StyledItem
-        data-cy="downshift-input-item"
-        data-testid="downshift-input-item"
-        className="dropdown-item"
-        css={{
-          backgroundColor: focused ? '$interaction' : '$bgDefault', width: '100%', padding: '0',
-        }}
-        isFocused={focused}
-      >
-        {type === 'color' && (
-          <StyledItemColorDiv>
-            <StyledItemColor style={{ backgroundColor: resolvedToken?.value.toString() }} />
-          </StyledItemColorDiv>
-        )}
-        <StyledItemName css={{ flexGrow: '1' }}>{getHighlightedText(resolvedToken?.name ?? '', search || '')}</StyledItemName>
-        <StyledItemValue>{resolvedToken?.value}</StyledItemValue>
-      </StyledItem>
-    );
-  }, [resolvedTokens, type]);
-
-  const renderDisplayTransform = React.useCallback((id: string, display: string) => `{${display}}`, []);
 
   return (
     <Downshift onSelect={handleSelect}>
@@ -369,26 +216,20 @@ export const DownshiftInput: React.FunctionComponent<DownShiftProps> = ({
           <Box css={{ display: 'flex', position: 'relative', width: '100%' }} className="input" ref={inputContainerRef}>
             {!!inlineLabel && !prefix && <Tooltip label={name}><StyledPrefix isText>{label}</StyledPrefix></Tooltip>}
             {!!prefix && <StyledPrefix>{prefix}</StyledPrefix>}
-            <MentionsInput
-              singleLine
-              style={{ ...mentionInputStyles }}
-              value={value}
-              onChange={handleMentionInputChange}
-              placeholder={placeholder}
-              autoComplete="off"
-              allowSpaceInQuery={false}
-              customSuggestionsContainer={renderMentionList}
-              onBlur={handleInputBlur}
+            <MentionInput
               name={name}
-            >
-              <Mention
-                trigger="{"
-                data={mentionData}
-                markup=" __id__ "
-                renderSuggestion={renderMentionListItem}
-                displayTransform={renderDisplayTransform}
-              />
-            </MentionsInput>
+              type={type}
+              value={value}
+              initialName={initialName}
+              placeholder={placeholder}
+              resolvedTokens={resolvedTokens}
+              inputContainerWith={inputContainerWith}
+              inputContainerPosX={inputContainerPosX}
+              inputContainerPosY={inputContainerPosY}
+              handleChange={handleChange}
+              handleBlur={handleBlur}
+              portalPlaceholder={portalPlaceholder}
+            />
             {suffix && (
               <StyledInputSuffix type="button" data-testid="downshift-input-suffix-button" onClick={handleAutoSuggest}>
                 <StyledIconDisclosure />
